First Commit
 Thursday 17/04/2025
This project is a web interface designed with a consistent and detailed pixel art style. The goal was to combine a nostalgic retro feel with modern web technologies to create an engaging and responsive experience. All visual elements were custom-made to match the pixel theme. The project highlights my skills in front-end development, creative design, and building user interfaces with attention to both function and aesthetics.